+++ .php-cs-fixer.php
@@ -1,9 +1,0 @@
-<?php
-
-$config = require __DIR__.'/vendor/contentful/core/scripts/php-cs-fixer.php';
-
-return $config(
-    'laravel',
-    true,
-    ['scripts', 'src', 'tests']
-);