+++ src/Facades/ContentfulCDA.php
@@ -1,0 +1,28 @@
+<?php
+
+/**
+ * This file is part of the contentful/laravel package.
+ *
+ * @copyright 2015-2024 Contentful GmbH
+ * @license   MIT
+ */
+
+declare(strict_types=1);
+
+namespace Presentcompany\ContentfulLaravel\Facades;
+
+use Contentful\Delivery\Client;
+use Illuminate\Support\Facades\Facade;
+
+class ContentfulCDA extends Facade
+{
+    /**
+     * Get the registered name of the component.
+     *
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return Client::class;
+    }
+}